+++ sdk/python/pulumi_aws/alb/target_group_attachment.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from .. import _utilities, _tables
 
 __all__ = ['TargetGroupAttachment']
@@ -16,7 +16,7 @@ class TargetGroupAttachment(pulumi.CustomResource):
                  resource_name: str,
                  opts: Optional[pulumi.ResourceOptions] = None,
                  availability_zone: Optional[pulumi.Input[str]] = None,
-                 port: Optional[pulumi.Input[float]] = None,
+                 port: Optional[pulumi.Input[int]] = None,
                  target_group_arn: Optional[pulumi.Input[str]] = None,
                  target_id: Optional[pulumi.Input[str]] = None,
                  __props__=None,
@@ -34,9 +34,9 @@ class TargetGroupAttachment(pulumi.CustomResource):
         import pulumi_aws as aws
 
         test_target_group = aws.lb.TargetGroup("testTargetGroup")
-        # Other arguments
+        # ... other configuration ...
         test_instance = aws.ec2.Instance("testInstance")
-        # Other arguments
+        # ... other configuration ...
         test_target_group_attachment = aws.lb.TargetGroupAttachment("testTargetGroupAttachment",
             target_group_arn=test_target_group.arn,
             target_id=test_instance.id,
@@ -50,7 +50,7 @@ class TargetGroupAttachment(pulumi.CustomResource):
 
         test_target_group = aws.lb.TargetGroup("testTargetGroup", target_type="lambda")
         test_function = aws.lambda_.Function("testFunction")
-        # Other arguments
+        # ... other configuration ...
         with_lb = aws.lambda_.Permission("withLb",
             action="lambda:InvokeFunction",
             function=test_function.arn,
@@ -65,7 +65,7 @@ class TargetGroupAttachment(pulumi.CustomResource):
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] availability_zone: The Availability Zone where the IP address of the target is to be registered. If the private ip address is outside of the VPC scope, this value must be set to 'all'.
-        :param pulumi.Input[float] port: The port on which targets receive traffic.
+        :param pulumi.Input[int] port: The port on which targets receive traffic.
         :param pulumi.Input[str] target_group_arn: The ARN of the target group with which to register targets
         :param pulumi.Input[str] target_id: The ID of the target. This is the Instance ID for an instance, or the container ID for an ECS container. If the target type is ip, specify an IP address. If the target type is lambda, specify the arn of lambda.
         """
@@ -107,7 +107,7 @@ class TargetGroupAttachment(pulumi.CustomResource):
             id: pulumi.Input[str],
             opts: Optional[pulumi.ResourceOptions] = None,
             availability_zone: Optional[pulumi.Input[str]] = None,
-            port: Optional[pulumi.Input[float]] = None,
+            port: Optional[pulumi.Input[int]] = None,
             target_group_arn: Optional[pulumi.Input[str]] = None,
             target_id: Optional[pulumi.Input[str]] = None) -> 'TargetGroupAttachment':
         """
@@ -118,7 +118,7 @@ class TargetGroupAttachment(pulumi.CustomResource):
         :param pulumi.Input[str] id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] availability_zone: The Availability Zone where the IP address of the target is to be registered. If the private ip address is outside of the VPC scope, this value must be set to 'all'.
-        :param pulumi.Input[float] port: The port on which targets receive traffic.
+        :param pulumi.Input[int] port: The port on which targets receive traffic.
         :param pulumi.Input[str] target_group_arn: The ARN of the target group with which to register targets
         :param pulumi.Input[str] target_id: The ID of the target. This is the Instance ID for an instance, or the container ID for an ECS container. If the target type is ip, specify an IP address. If the target type is lambda, specify the arn of lambda.
         """
@@ -142,7 +142,7 @@ class TargetGroupAttachment(pulumi.CustomResource):
 
     @property
     @pulumi.getter
-    def port(self) -> pulumi.Output[Optional[float]]:
+    def port(self) -> pulumi.Output[Optional[int]]:
         """
         The port on which targets receive traffic.
         """
