+++ sdk/python/pulumi_aws/directconnect/bgp_peer.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from .. import _utilities, _tables
 
 __all__ = ['BgpPeer']
@@ -17,7 +17,7 @@ class BgpPeer(pulumi.CustomResource):
                  opts: Optional[pulumi.ResourceOptions] = None,
                  address_family: Optional[pulumi.Input[str]] = None,
                  amazon_address: Optional[pulumi.Input[str]] = None,
-                 bgp_asn: Optional[pulumi.Input[float]] = None,
+                 bgp_asn: Optional[pulumi.Input[int]] = None,
                  bgp_auth_key: Optional[pulumi.Input[str]] = None,
                  customer_address: Optional[pulumi.Input[str]] = None,
                  virtual_interface_id: Optional[pulumi.Input[str]] = None,
@@ -44,7 +44,7 @@ class BgpPeer(pulumi.CustomResource):
         :param pulumi.Input[str] address_family: The address family for the BGP peer. `ipv4 ` or `ipv6`.
         :param pulumi.Input[str] amazon_address: The IPv4 CIDR address to use to send traffic to Amazon.
                Required for IPv4 BGP peers on public virtual interfaces.
-        :param pulumi.Input[float] bgp_asn: The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.
+        :param pulumi.Input[int] bgp_asn: The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.
         :param pulumi.Input[str] bgp_auth_key: The authentication key for BGP configuration.
         :param pulumi.Input[str] customer_address: The IPv4 CIDR destination address to which Amazon should send traffic.
                Required for IPv4 BGP peers on public virtual interfaces.
@@ -95,7 +95,7 @@ class BgpPeer(pulumi.CustomResource):
             address_family: Optional[pulumi.Input[str]] = None,
             amazon_address: Optional[pulumi.Input[str]] = None,
             aws_device: Optional[pulumi.Input[str]] = None,
-            bgp_asn: Optional[pulumi.Input[float]] = None,
+            bgp_asn: Optional[pulumi.Input[int]] = None,
             bgp_auth_key: Optional[pulumi.Input[str]] = None,
             bgp_peer_id: Optional[pulumi.Input[str]] = None,
             bgp_status: Optional[pulumi.Input[str]] = None,
@@ -112,7 +112,7 @@ class BgpPeer(pulumi.CustomResource):
         :param pulumi.Input[str] amazon_address: The IPv4 CIDR address to use to send traffic to Amazon.
                Required for IPv4 BGP peers on public virtual interfaces.
         :param pulumi.Input[str] aws_device: The Direct Connect endpoint on which the BGP peer terminates.
-        :param pulumi.Input[float] bgp_asn: The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.
+        :param pulumi.Input[int] bgp_asn: The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.
         :param pulumi.Input[str] bgp_auth_key: The authentication key for BGP configuration.
         :param pulumi.Input[str] bgp_peer_id: The ID of the BGP peer.
         :param pulumi.Input[str] bgp_status: The Up/Down state of the BGP peer.
@@ -162,7 +162,7 @@ class BgpPeer(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="bgpAsn")
-    def bgp_asn(self) -> pulumi.Output[float]:
+    def bgp_asn(self) -> pulumi.Output[int]:
         """
         The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.
         """
